remove IntegratorHPMCMonoImplicit class
unify with IntegratorHPMCMono
