+++ hoomd/hpmc/IntegratorHPMCMono.h
@@ -23,6 +23,10 @@
 #include "hoomd/RNGIdentifiers.h"
 #include "hoomd/managed_allocator.h"
 
+#ifdef ENABLE_TBB
+#include <thread>
+#endif
+
 #ifdef ENABLE_MPI
 #include "hoomd/Communicator.h"
 #include "hoomd/HOOMDMPI.h"
@@ -139,6 +143,62 @@ class IntegratorHPMCMono : public IntegratorHPMC
 
         //! Take one timestep forward
         virtual void update(unsigned int timestep);
+
+        /*
+         * Depletant related options
+         */
+
+        //! Set the depletant density in the free volume
+        void setDepletantFugacity(unsigned int type, Scalar fugacity)
+            {
+            if (type >= this->m_pdata->getNTypes())
+                throw std::runtime_error("Unknown type.");
+            m_fugacity[type] = fugacity;
+            }
+
+        //! Returns the depletant fugacity
+        Scalar getDepletantFugacity(unsigned int type)
+            {
+            return m_fugacity[type];
+            }
+
+        //! Set quermass integration mode
+        void setQuermassMode(bool enable_quermass)
+            {
+            if (enable_quermass && !Shape::supportsSweepRadius())
+                throw std::runtime_error("Quermass integration not supported for this shape");
+
+            m_quermass = enable_quermass;
+            }
+
+        //! Get the quermass integration state
+        bool getQuermassMode()
+            {
+            return m_quermass;
+            }
+
+        //! Set up the additional sweep radius around every shape
+        void setSweepRadius(Scalar sweep_radius)
+            {
+            // check if supported
+            if (sweep_radius != 0.0 && !Shape::supportsSweepRadius())
+                throw std::runtime_error("This shape doesn's support setting a sweep radius to extend the surface out.\n");
+
+            m_sweep_radius = sweep_radius;
+            }
+
+        //! Get the sweep radius
+        Scalar getSweepRadius()
+             {
+             return m_sweep_radius;
+             }
+
+        //! Get the current counter values
+        std::vector<hpmc_implicit_counters_t> getImplicitCounters(unsigned int mode=0);
+
+        /*
+         * Common HPMC API
+         */
 
         //! Get the maximum particle diameter
         virtual Scalar getMaxCoreDiameter();
@@ -346,6 +406,30 @@ class IntegratorHPMCMono : public IntegratorHPMC
 
         Index2D m_overlap_idx;                      //!!< Indexer for interaction matrix
 
+        /* Depletants related data members */
+
+        std::vector<Scalar> m_fugacity;                          //!< Average depletant number density in free volume, per type
+
+        GlobalArray<hpmc_implicit_counters_t> m_implicit_count;               //!< Counter of depletant insertions
+        std::vector<hpmc_implicit_counters_t> m_implicit_count_run_start;     //!< Counter of depletant insertions at run start
+        std::vector<hpmc_implicit_counters_t> m_implicit_count_step_start;    //!< Counter of depletant insertions at step start
+
+        bool m_quermass;                                         //!< True if quermass integration mode is enabled
+        Scalar m_sweep_radius;                                   //!< Radius of sphere to sweep shapes by
+
+        //! Test whether to reject the current particle move based on depletants
+        #ifndef ENABLE_TBB
+        inline bool checkDepletantOverlap(unsigned int i, vec3<Scalar> pos_i, Shape shape_i, unsigned int typ_i,
+            Scalar4 *h_postype, Scalar4 *h_orientation, unsigned int *h_overlaps,
+            hpmc_counters_t& counters, hpmc_implicit_counters_t *implicit_counters,
+            hoomd::RandomGenerator& rng_depletants);
+        #else
+        inline bool checkDepletantOverlap(unsigned int i, vec3<Scalar> pos_i, Shape shape_i, unsigned int typ_i,
+            Scalar4 *h_postype, Scalar4 *h_orientation, unsigned int *h_overlaps,
+            hpmc_counters_t& counters, hpmc_implicit_counters_t *implicit_counters,
+            tbb::enumerable_thread_specific< hoomd::RandomGenerator >& rng_depletants_parallel);
+        #endif
+
         //! Set the nominal width appropriate for looped moves
         virtual void updateCellWidth();
 
@@ -380,7 +464,9 @@ IntegratorHPMCMono<Shape>::IntegratorHPMCMono(std::shared_ptr<SystemDefinition> 
               m_image_list_is_initialized(false),
               m_image_list_valid(false),
               m_hasOrientation(true),
-              m_extra_image_width(0.0)
+              m_extra_image_width(0.0),
+              m_quermass(false),
+              m_sweep_radius(0.0)
     {
     // allocate the parameter storage
     m_params = std::vector<param_type, managed_allocator<param_type> >(m_pdata->getNTypes(), param_type(), managed_allocator<param_type>(m_exec_conf->isCUDAEnabled()));
@@ -401,8 +487,59 @@ IntegratorHPMCMono<Shape>::IntegratorHPMCMono(std::shared_ptr<SystemDefinition> 
     m_aabbs = NULL;
     m_aabbs_capacity = 0;
     m_aabb_tree_invalid = true;
+
+    GlobalArray<hpmc_implicit_counters_t> implicit_count(this->m_pdata->getNTypes(),this->m_exec_conf);
+    m_implicit_count.swap(implicit_count);
+    TAG_ALLOCATION(m_implicit_count);
+
+        {
+        ArrayHandle<hpmc_implicit_counters_t> h_implicit_count(m_implicit_count, access_location::host, access_mode::readwrite);
+        for (unsigned int i = 0; i < this->m_pdata->getNTypes(); ++i)
+            h_implicit_count.data[i] = hpmc_implicit_counters_t();
+        }
+
+    m_implicit_count_run_start.resize(this->m_pdata->getNTypes());
+    m_implicit_count_step_start.resize(this->m_pdata->getNTypes());
+
+    m_fugacity.resize(this->m_pdata->getNTypes(),0.0);
     }
 
+/*! \param mode 0 -> Absolute count, 1 -> relative to the start of the run, 2 -> relative to the last executed step
+    \return The current state of the acceptance counters
+
+    IntegratorHPMCMonoImplicit maintains a count of the number of accepted and rejected moves since instantiation. getCounters()
+    provides the current value. The parameter *mode* controls whether the returned counts are absolute, relative
+    to the start of the run, or relative to the start of the last executed step.
+*/
+template<class Shape>
+std::vector<hpmc_implicit_counters_t> IntegratorHPMCMono<Shape>::getImplicitCounters(unsigned int mode)
+    {
+    ArrayHandle<hpmc_implicit_counters_t> h_counters(m_implicit_count, access_location::host, access_mode::read);
+    std::vector<hpmc_implicit_counters_t> result(this->m_pdata->getNTypes());
+
+    std::copy(h_counters.data, h_counters.data + this->m_pdata->getNTypes(), result.begin());
+    if (mode == 1)
+        {
+        for (unsigned int i = 0; i < this->m_pdata->getNTypes(); ++i)
+            result[i] = result[i] - m_implicit_count_run_start[i];
+        }
+    else
+        {
+        for (unsigned int i = 0; i < this->m_pdata->getNTypes(); ++i)
+            result[i] = result[i] - m_implicit_count_step_start[i];
+        }
+
+    #ifdef ENABLE_MPI
+    if (this->m_comm)
+        {
+        // MPI Reduction to total result values on all ranks
+        for (unsigned int i = 0; i < this->m_pdata->getNTypes(); ++i)
+            MPI_Allreduce(MPI_IN_PLACE, &result[i].insert_count, 1, MPI_LONG_LONG_INT, MPI_SUM, this->m_exec_conf->getMPICommunicator());
+        }
+    #endif
+
+    return result;
+    }
 
 template<class Shape>
 std::vector< std::string > IntegratorHPMCMono<Shape>::getProvidedLogQuantities()
@@ -415,6 +552,15 @@ std::vector< std::string > IntegratorHPMCMono<Shape>::getProvidedLogQuantities()
         result.push_back("hpmc_patch_energy");
         result.push_back("hpmc_patch_rcut");
         }
+
+    for (unsigned int typ=0; typ<this->m_pdata->getNTypes();typ++)
+      {
+      std::ostringstream tmp_str0;
+      tmp_str0<<"hpmc_fugacity_"<< this->m_pdata->getNameByType(typ);
+      result.push_back(tmp_str0.str());
+      }
+
+    result.push_back("hpmc_insert_count");
 
     return result;
     }
@@ -446,17 +592,78 @@ Scalar IntegratorHPMCMono<Shape>::getLogValue(const std::string& quantity, unsig
             throw std::runtime_error("Error getting log value");
             }
         }
-    else
+
+    //loop over per particle fugacities
+    for (unsigned int typ=0; typ<this->m_pdata->getNTypes();typ++)
         {
-        //nothing found -> pass on to integrator
-        return IntegratorHPMC::getLogValue(quantity, timestep);
+        std::ostringstream tmp_str0;
+        tmp_str0<<"hpmc_fugacity_"<<this->m_pdata->getNameByType(typ);
+        if (quantity==tmp_str0.str())
+            return m_fugacity[typ];
         }
+
+    hpmc_counters_t counters = IntegratorHPMC::getCounters(2);
+    const std::vector<hpmc_implicit_counters_t>& implicit_counters = getImplicitCounters(2);
+
+    if (quantity == "hpmc_insert_count")
+        {
+        // reduce over all types
+        unsigned long long int total_insert_count = 0;
+        for (unsigned int i = 0; i < this->m_pdata->getNTypes(); ++i)
+            total_insert_count += implicit_counters[i].insert_count;
+
+        // return number of depletant insertions per colloid
+        if (counters.getNMoves() > 0)
+            return (Scalar)total_insert_count/(Scalar)counters.getNMoves();
+        else
+            return Scalar(0.0);
+        }
+
+    //nothing found -> pass on to base class
+    return IntegratorHPMC::getLogValue(quantity, timestep);
     }
 
 template <class Shape>
 void IntegratorHPMCMono<Shape>::printStats()
     {
     IntegratorHPMC::printStats();
+
+    const std::vector<hpmc_implicit_counters_t>& result = getImplicitCounters(1);
+    hpmc_counters_t counters = getCounters(1);
+
+    // reduce over all types
+    unsigned long long int total_insert_count = 0;
+    for (unsigned int i = 0; i < this->m_pdata->getNTypes(); ++i)
+        total_insert_count += result[i].insert_count;
+
+    double cur_time = double(this->m_clock.getTime()) / Scalar(1e9);
+
+    bool has_depletants = false;
+    for (unsigned int i = 0; i < this->m_pdata->getNTypes(); ++i)
+        {
+        if (m_fugacity[i] != 0.0)
+            {
+            has_depletants = true;
+            break;
+            }
+        }
+
+    if (!has_depletants)
+        return;
+
+    this->m_exec_conf->msg->notice(2) << "-- Implicit depletants stats:" << "\n";
+    this->m_exec_conf->msg->notice(2) << "Depletant insertions per trial move:      "
+        << double(total_insert_count)/cur_time << "\n";
+
+    // supply additional statistics
+    for (unsigned int i = 0; i < this->m_pdata->getNTypes(); ++i)
+        {
+        if (m_fugacity[i] != 0.0)
+            {
+            this->m_exec_conf->msg->notice(3) << "Type '" << this->m_pdata->getNameByType(i) << "': "
+                << double(result[i].insert_count)/double(counters.getNMoves()) << std::endl;
+            }
+        }
 
     /*unsigned int max_height = 0;
     unsigned int total_height = 0;
@@ -477,6 +684,10 @@ template <class Shape>
 void IntegratorHPMCMono<Shape>::resetStats()
     {
     IntegratorHPMC::resetStats();
+
+    ArrayHandle<hpmc_implicit_counters_t> h_counters(m_implicit_count, access_location::host, access_mode::read);
+    for (unsigned int i = 0; i < this->m_pdata->getNTypes(); ++i)
+        m_implicit_count_run_start[i] = h_counters.data[i];
     }
 
 template <class Shape>
@@ -501,6 +712,21 @@ void IntegratorHPMCMono<Shape>::slotNumTypesChange()
     m_overlaps.swap(overlaps);
 
     updateCellWidth();
+
+    // depletant related counters
+    unsigned int old_ntypes = m_implicit_count.getNumElements();
+    m_implicit_count.resize(this->m_pdata->getNTypes());
+
+        {
+        ArrayHandle<hpmc_implicit_counters_t> h_implicit_count(m_implicit_count, access_location::host, access_mode::readwrite);
+        for (unsigned int i = old_ntypes; i < this->m_pdata->getNTypes(); ++i)
+            h_implicit_count.data[i] = hpmc_implicit_counters_t();
+        }
+    m_implicit_count_run_start.resize(this->m_pdata->getNTypes());
+    m_implicit_count_step_start.resize(this->m_pdata->getNTypes());
+
+    // depletant fugacities
+    m_fugacity.resize(this->m_pdata->getNTypes(),0.0);
     }
 
 template <class Shape>
@@ -512,6 +738,10 @@ void IntegratorHPMCMono<Shape>::update(unsigned int timestep)
     // get needed vars
     ArrayHandle<hpmc_counters_t> h_counters(m_count_total, access_location::host, access_mode::readwrite);
     hpmc_counters_t& counters = h_counters.data[0];
+
+    ArrayHandle<hpmc_implicit_counters_t> h_implicit_counters(m_implicit_count, access_location::host, access_mode::readwrite);
+    std::copy(h_implicit_counters.data, h_implicit_counters.data + this->m_pdata->getNTypes(), m_implicit_count_step_start.begin());
+
     const BoxDim& box = m_pdata->getBox();
     unsigned int ndim = this->m_sysdef->getNDimensions();
 
@@ -531,6 +761,26 @@ void IntegratorHPMCMono<Shape>::update(unsigned int timestep)
     limitMoveDistances();
     // update the image list
     updateImageList();
+
+    // Combine the three seeds to generate RNG for poisson distribution
+    #ifndef ENABLE_TBB
+    hoomd::RandomGenerator rng_depletants(this->m_seed,
+        timestep,
+        this->m_exec_conf->getRank(),
+        hoomd::RNGIdentifier::HPMCDepletants);
+    #else
+    // create one RNG per thread
+    tbb::enumerable_thread_specific< hoomd::RandomGenerator > rng_depletants_parallel([=]
+        {
+        std::vector<unsigned int> seed_seq(5);
+        std::hash<std::thread::id> hash;
+        return hoomd::RandomGenerator(this->m_seed,
+            timestep,
+            this->m_exec_conf->getRank(),
+            hash(std::this_thread::get_id()),
+            hoomd::RNGIdentifier::HPMCDepletants);
+        });
+    #endif
 
     if (this->m_prof) this->m_prof->push(this->m_exec_conf, "HPMC update");
 
@@ -810,9 +1060,21 @@ void IntegratorHPMCMono<Shape>::update(unsigned int timestep)
                 patch_field_energy_diff -= m_external->energydiff(i, pos_old, shape_old, pos_i, shape_i);
                 }
 
+            bool accept = !overlap && hoomd::detail::generate_canonical<double>(rng_i) < slow::exp(patch_field_energy_diff);
+
+            // The trial move is valid, so check if it is invalidated by depletants
+            if (accept)
+                {
+                #ifndef ENABLE_TBB
+                accept = checkDepletantOverlap(i, pos_i, shape_i, typ_i, h_postype.data, h_orientation.data, h_overlaps.data, counters, h_implicit_counters.data, rng_depletants);
+                #else
+                accept = checkDepletantOverlap(i, pos_i, shape_i, typ_i, h_postype.data, h_orientation.data, h_overlaps.data, counters, h_implicit_counters.data, rng_depletants_parallel);
+                #endif
+                }
+
             // If no overlaps and Metropolis criterion is met, accept
             // trial move and update positions  and/or orientations.
-            if (!overlap && hoomd::detail::generate_canonical<double>(rng_i) < slow::exp(patch_field_energy_diff))
+            if (accept)
                 {
                 // increment accept counter and assign new position
                 if (!shape_i.ignoreStatistics())
@@ -1437,23 +1699,42 @@ inline const std::vector<vec3<Scalar> >& IntegratorHPMCMono<Shape>::updateImageL
 template <class Shape>
 void IntegratorHPMCMono<Shape>::updateCellWidth()
     {
-    m_nominal_width = getMaxCoreDiameter();
+    this->m_nominal_width = this->getMaxCoreDiameter();
 
-    if (m_patch)
+    Scalar max_d(0.0);
+
+    for (unsigned int type = 0; type < this->m_pdata->getNTypes(); ++type)
+        {
+        if (m_fugacity[type] != Scalar(0.0))
+            {
+            // add range of depletion interaction
+            quat<Scalar> o;
+            Shape tmp(o, this->m_params[type]);
+            max_d = std::max(max_d, (Scalar) tmp.getCircumsphereDiameter());
+            }
+        }
+
+    // extend the image list by the depletant diameter, since we're querying
+    // AABBs that are larger than the shape diameters themselves
+    this->m_extra_image_width = m_quermass ? 2.0*m_sweep_radius : max_d;
+
+    this->m_nominal_width += this->m_extra_image_width;
+
+    // Account for patch width
+    if (this->m_patch)
         {
         Scalar max_extent = 0.0;
         for (unsigned int typ = 0; typ < this->m_pdata->getNTypes(); typ++)
             {
-            max_extent = std::max(max_extent, m_patch->getAdditiveCutoff(typ));
+            max_extent = std::max(max_extent, this->m_patch->getAdditiveCutoff(typ));
             }
 
-        m_nominal_width = std::max(m_nominal_width, max_extent+m_patch->getRCut());
+        this->m_nominal_width = std::max(this->m_nominal_width, this->m_patch->getRCut() + max_extent);
         }
+    this->m_image_list_valid = false;
+    this->m_aabb_tree_invalid = true;
 
-    // changing the cell width means that the particle shapes have changed, assume this invalidates the
-    // image list and aabb tree
-    m_image_list_valid = false;
-    m_aabb_tree_invalid = true;
+    this->m_exec_conf->msg->notice(5) << "IntegratorHPMCMonoImplicit: updating nominal width to " << this->m_nominal_width << std::endl;
     }
 
 template <class Shape>
@@ -1804,6 +2085,889 @@ bool IntegratorHPMCMono<Shape>::py_test_overlap(unsigned int type_i, unsigned in
     return overlap;
     }
 
+/*! \param i The particle id in the list
+    \param pos_i Particle position being tested
+    \param shape_i Particle shape (including orientation) being tested
+    \param typ_i Type of the particle being tested
+    \param h_postype Pointer to GPUArray containing particle positions
+    \param h_orientation Pointer to GPUArray containing particle orientations
+    \param h_overlaps Pointer to GPUArray containing interaction matrix
+    \param hpmc_counters_t&  Pointer to current counters
+    \param hpmc_implicit_counters_t&  Pointer to current implicit counters
+    \param rng_depletants The RNG used within this algorithm
+    \param rng_i The RNG used for evaluating the Metropolis criterion
+
+    In order to determine whether or not moves are accepted, particle positions are checked against a randomly generated set of depletant positions.
+
+    NOTE: To avoid numerous acquires and releases of GPUArrays, data pointers are passed directly into this const function.
+    */
+#ifndef ENABLE_TBB
+template<class Shape>
+inline bool IntegratorHPMCMono<Shape>::checkDepletantOverlap(unsigned int i, vec3<Scalar> pos_i,
+    Shape shape_i, unsigned int typ_i, Scalar4 *h_postype, Scalar4 *h_orientation, unsigned int *h_overlaps,
+    hpmc_counters_t& counters, hpmc_implicit_counters_t *implicit_counters, hoomd::RandomGenerator& rng_depletants)
+#else
+template<class Shape>
+inline bool IntegratorHPMCMono<Shape>::checkDepletantOverlap(unsigned int i, vec3<Scalar> pos_i,
+    Shape shape_i, unsigned int typ_i, Scalar4 *h_postype, Scalar4 *h_orientation, unsigned int *h_overlaps,
+    hpmc_counters_t& counters, hpmc_implicit_counters_t *implicit_counters,
+    tbb::enumerable_thread_specific< hoomd::RandomGenerator >& rng_depletants_parallel)
+#endif
+    {
+    bool accept = true;
+
+    const unsigned int n_images = this->m_image_list.size();
+
+    Shape shape_old(quat<Scalar>(h_orientation[i]), this->m_params[typ_i]);
+    detail::AABB aabb_i_local = shape_i.getAABB(vec3<Scalar>(0,0,0));
+    detail::AABB aabb_i_local_old = shape_old.getAABB(vec3<Scalar>(0,0,0));
+
+    #ifdef ENABLE_TBB
+    std::vector< tbb::enumerable_thread_specific<hpmc_implicit_counters_t> > thread_implicit_counters(this->m_pdata->getNTypes());
+    tbb::enumerable_thread_specific<hpmc_counters_t> thread_counters;
+    #endif
+
+    #ifdef ENABLE_TBB
+    try {
+    #endif
+    for (unsigned int type = 0; type < this->m_pdata->getNTypes(); ++type)
+        {
+        if (m_fugacity[type] == 0.0 || (!h_overlaps[this->m_overlap_idx(type, typ_i)] && !m_quermass))
+            continue;
+
+        std::vector<unsigned int> intersect_i;
+        std::vector<unsigned int> image_i;
+        std::vector<detail::AABB> aabbs_i;
+
+        if (accept && m_fugacity[type] > 0.0)
+            {
+            // find neighbors whose circumspheres overlap particle i's circumsphere in the old configuration
+            // Here, circumsphere refers to the sphere around the depletant-excluded volume
+
+            Shape tmp(quat<Scalar>(), this->m_params[type]);
+            Scalar d_dep = tmp.getCircumsphereDiameter();
+
+            Scalar range = m_quermass ? Scalar(2.0)*m_sweep_radius : d_dep;
+
+            // get old AABB and extend
+            vec3<Scalar> lower = aabb_i_local_old.getLower();
+            vec3<Scalar> upper = aabb_i_local_old.getUpper();
+            lower.x -= range; lower.y -= range; lower.z -= range;
+            upper.x += range; upper.y += range; upper.z += range;
+            detail::AABB aabb_local = detail::AABB(lower,upper);
+
+            vec3<Scalar> pos_i_old(h_postype[i]);
+
+            // All image boxes (including the primary)
+            for (unsigned int cur_image = 0; cur_image < n_images; cur_image++)
+                {
+                vec3<Scalar> pos_i_old_image = pos_i_old + this->m_image_list[cur_image];
+                detail::AABB aabb = aabb_local;
+                aabb.translate(pos_i_old_image);
+
+                // stackless search
+                for (unsigned int cur_node_idx = 0; cur_node_idx < this->m_aabb_tree.getNumNodes(); cur_node_idx++)
+                    {
+                    if (detail::overlap(this->m_aabb_tree.getNodeAABB(cur_node_idx), aabb))
+                        {
+                        if (this->m_aabb_tree.isNodeLeaf(cur_node_idx))
+                            {
+                            for (unsigned int cur_p = 0; cur_p < this->m_aabb_tree.getNodeNumParticles(cur_node_idx); cur_p++)
+                                {
+                                // read in its position and orientation
+                                unsigned int j = this->m_aabb_tree.getNodeParticle(cur_node_idx, cur_p);
+
+                                if (i == j && cur_image == 0) continue;
+
+                                // load the old position and orientation of the j particle
+                                Scalar4 postype_j = h_postype[j];
+                                vec3<Scalar> pos_j(postype_j);
+                                vec3<Scalar> r_ij = pos_j - pos_i_old_image;
+
+                                unsigned int typ_j = __scalar_as_int(postype_j.w);
+                                Shape shape_j(quat<Scalar>(), this->m_params[typ_j]);
+                                if (shape_j.hasOrientation())
+                                    shape_j.orientation = quat<Scalar>(h_orientation[j]);
+
+                                detail::AABB aabb_j = shape_j.getAABB(pos_j);
+
+                                // check AABB overlap
+                                bool overlap_excluded = detail::overlap(aabb, aabb_j);
+
+                                // currently ignore overlap flags with quermass==true
+                                overlap_excluded = overlap_excluded && (m_quermass || h_overlaps[this->m_overlap_idx(type,typ_j)]);
+
+                                if (Shape::supportsSweepRadius() && overlap_excluded)
+                                    {
+                                    // do i and j overlap with their excluded volumes in the old configuration?
+                                    #ifdef ENABLE_TBB
+                                    thread_counters.local().overlap_checks++;
+                                    #else
+                                    counters.overlap_checks++;
+                                    #endif
+
+                                    unsigned int err = 0;
+                                    Scalar sweep_radius = m_quermass ? m_sweep_radius : Scalar(0.5)*tmp.getCircumsphereDiameter();
+                                    overlap_excluded = test_overlap(r_ij, shape_old, shape_j, err, sweep_radius, sweep_radius);
+
+                                    if (err)
+                                    #ifdef ENABLE_TBB
+                                        thread_counters.local().overlap_err_count++;
+                                    #else
+                                        counters.overlap_err_count++;
+                                    #endif
+                                    }
+                                if (overlap_excluded)
+                                    {
+                                    intersect_i.push_back(j);
+                                    image_i.push_back(cur_image);
+
+                                    // cache the translated AABB of particle j. If i's image is cur_image, then j's
+                                    // image is the negative of that (and we use i's untranslated position below)
+                                    aabb_j.translate(-this->m_image_list[cur_image]);
+                                    aabbs_i.push_back(aabb_j);
+                                    }
+                                }
+                            }
+                        }
+                    else
+                        {
+                        // skip ahead
+                        cur_node_idx += this->m_aabb_tree.getNodeSkip(cur_node_idx);
+                        }
+                    }  // end loop over AABB nodes
+                } // end loop over images
+
+            // now, we have a list of intersecting spheres, sample in the union of intersection volumes
+            // we sample from their union by checking if any generated position falls in the intersection
+            // of their AABBs, only accepting it if it was generated from neighbor j_min
+
+            // world AABB of particle i
+            detail::AABB aabb_i_old = aabb_i_local_old;
+            aabb_i_old.translate(pos_i_old);
+
+            // extend AABB i by sweep radius
+            range = Scalar(0.5)*d_dep+m_sweep_radius;
+            vec3<Scalar> lower_i = aabb_i_old.getLower();
+            vec3<Scalar> upper_i = aabb_i_old.getUpper();
+            lower_i.x -= range; lower_i.y -= range; lower_i.z -= range;
+            upper_i.x += range; upper_i.y += range; upper_i.z += range;
+
+            // for every pairwise intersection
+            #ifdef ENABLE_TBB
+            tbb::parallel_for(tbb::blocked_range<unsigned int>(0, (unsigned int)intersect_i.size()),
+                [=, &intersect_i, &image_i, &aabbs_i,
+                    &accept, &rng_depletants_parallel,
+                    &thread_counters, &thread_implicit_counters](const tbb::blocked_range<unsigned int>& s) {
+            for (unsigned int k = s.begin(); k != s.end(); ++k)
+            #else
+            for (unsigned int k = 0; k < intersect_i.size(); ++k)
+            #endif
+                {
+                // world AABB of shape j, in image of i
+                detail::AABB aabb_j = aabbs_i[k];
+
+                // extend AABB j by sweep radius
+                vec3<Scalar> lower_j = aabb_j.getLower();
+                vec3<Scalar> upper_j = aabb_j.getUpper();
+                lower_j.x -= range; lower_j.y -= range; lower_j.z -= range;
+                upper_j.x += range; upper_j.y += range; upper_j.z += range;
+
+                // we already know the AABBs are overlapping, compute their intersection
+                vec3<Scalar> intersect_lower, intersect_upper;
+                intersect_lower.x = std::max(lower_i.x, lower_j.x);
+                intersect_lower.y = std::max(lower_i.y, lower_j.y);
+                intersect_lower.z = std::max(lower_i.z, lower_j.z);
+                intersect_upper.x = std::min(upper_i.x, upper_j.x);
+                intersect_upper.y = std::min(upper_i.y, upper_j.y);
+                intersect_upper.z = std::min(upper_i.z, upper_j.z);
+
+                detail::AABB aabb_intersect(intersect_lower, intersect_upper);
+
+                // intersectionAABB volume
+                Scalar V =  (intersect_upper.x-intersect_lower.x)*(intersect_upper.y-intersect_lower.y)*(intersect_upper.z-intersect_lower.z);
+
+                // chooose the number of depletants in the intersection volume
+                hoomd::PoissonDistribution<Scalar> poisson(m_fugacity[type]*V);
+                #ifdef ENABLE_TBB
+                hoomd::RandomGenerator& my_rng = rng_depletants_parallel.local();
+                #else
+                hoomd::RandomGenerator& my_rng = rng_depletants;
+                #endif
+
+                unsigned int n = poisson(my_rng);
+
+                // for every depletant
+                #ifdef ENABLE_TBB
+                tbb::parallel_for(tbb::blocked_range<unsigned int>(0, (unsigned int)n),
+                    [=, &intersect_i, &image_i, &aabbs_i,
+                        &accept, &rng_depletants_parallel,
+                        &thread_counters, &thread_implicit_counters](const tbb::blocked_range<unsigned int>& t) {
+                for (unsigned int l = t.begin(); l != t.end(); ++l)
+                #else
+                for (unsigned int l = 0; l < n; ++l)
+                #endif
+                    {
+                    #ifdef ENABLE_TBB
+                    hoomd::RandomGenerator& my_rng = rng_depletants_parallel.local();
+                    #else
+                    hoomd::RandomGenerator& my_rng = rng_depletants;
+                    #endif
+
+                    #ifdef ENABLE_TBB
+                    thread_implicit_counters[type].local().insert_count++;
+                    #else
+                    implicit_counters[type].insert_count++;
+                    #endif
+
+                    vec3<Scalar> pos_test = generatePositionInAABB(my_rng, aabb_intersect);
+                    Shape shape_test(quat<Scalar>(), this->m_params[type]);
+                    if (shape_test.hasOrientation())
+                        {
+                        shape_test.orientation = generateRandomOrientation(my_rng);
+                        }
+
+                    // check if depletant falls in other intersection volumes
+
+                    // load the depletant's AABB and extend it
+                    detail::AABB aabb_test = shape_test.getAABB(pos_test);
+                    vec3<Scalar> lower_test = aabb_test.getLower();
+                    vec3<Scalar> upper_test = aabb_test.getUpper();
+                    lower_test.x -= m_sweep_radius; lower_test.y -= m_sweep_radius; lower_test.z -= m_sweep_radius;
+                    upper_test.x += m_sweep_radius; upper_test.y += m_sweep_radius; upper_test.z += m_sweep_radius;
+                    aabb_test = detail::AABB(lower_test, upper_test);
+
+                    bool active = true;
+                    for (unsigned int m = 0; m < k; ++m)
+                        {
+                        if (detail::overlap(aabb_test,aabbs_i[m]))
+                            {
+                            active = false;
+                            break;
+                            }
+                        }
+
+                    if (!active)
+                        continue;
+
+                    // depletant falls in intersection volume between circumspheres
+
+                    if (!m_quermass)
+                        {
+                        // Check if the old configuration of particle i generates an overlap
+                        bool overlap_old = false;
+                            {
+                            vec3<Scalar> r_ij = pos_i_old - pos_test;
+                            OverlapReal rsq = dot(r_ij,r_ij);
+                            OverlapReal DaDb = shape_test.getCircumsphereDiameter() + shape_old.getCircumsphereDiameter() + Scalar(2.0)*m_sweep_radius;
+                            bool circumsphere_overlap = (rsq*OverlapReal(4.0) <= DaDb * DaDb);
+
+                            if (m_quermass || h_overlaps[this->m_overlap_idx(type, typ_i)])
+                                {
+                                #ifdef ENABLE_TBB
+                                thread_counters.local().overlap_checks++;
+                                #else
+                                counters.overlap_checks++;
+                                #endif
+
+                                unsigned int err = 0;
+                                if (circumsphere_overlap && test_overlap(r_ij, shape_test, shape_old, err, 0.0, m_sweep_radius))
+                                    {
+                                    overlap_old = true;
+                                    }
+                                if (err)
+                                #ifdef ENABLE_TBB
+                                    thread_counters.local().overlap_err_count++;
+                                #else
+                                    counters.overlap_err_count++;
+                                #endif
+                                }
+                            }
+
+                        // if not intersecting ptl i in old config, ignore
+                        if (!overlap_old)
+                            continue;
+
+                        // Check if the new configuration of particle i generates an overlap
+                        bool overlap_new = false;
+
+                            {
+                            vec3<Scalar> r_ij = pos_i - pos_test;
+
+                            OverlapReal rsq = dot(r_ij,r_ij);
+                            OverlapReal DaDb = shape_test.getCircumsphereDiameter() + shape_i.getCircumsphereDiameter() + Scalar(2.0)*m_sweep_radius;
+                            bool circumsphere_overlap = (rsq*OverlapReal(4.0) <= DaDb * DaDb);
+
+                            if (m_quermass || h_overlaps[this->m_overlap_idx(type, typ_i)])
+                                {
+                                #ifdef ENABLE_TBB
+                                thread_counters.local().overlap_checks++;
+                                #else
+                                counters.overlap_checks++;
+                                #endif
+
+                                unsigned int err = 0;
+                                if (circumsphere_overlap && test_overlap(r_ij, shape_test, shape_i, err, 0.0, m_sweep_radius))
+                                    {
+                                    overlap_new = true;
+                                    }
+                                if (err)
+                                #ifdef ENABLE_TBB
+                                    thread_counters.local().overlap_err_count++;
+                                #else
+                                    counters.overlap_err_count++;
+                                #endif
+                                }
+                            }
+
+                        if (overlap_new)
+                            continue;
+                        }
+
+                    // does the depletant fall into the overlap volume with other particles?
+                    bool in_intersection_volume = false;
+                    for (unsigned int m = 0; m < intersect_i.size(); ++m)
+                        {
+                        // read in its position and orientation
+                        unsigned int j = intersect_i[m];
+
+                        // load the old position and orientation of the j particle
+                        Scalar4 postype_j = h_postype[j];
+                        Scalar4 orientation_j = h_orientation[j];
+
+                        vec3<Scalar> r_jk = vec3<Scalar>(postype_j) - pos_test - this->m_image_list[image_i[m]];
+
+                        unsigned int typ_j = __scalar_as_int(postype_j.w);
+                        Shape shape_j(quat<Scalar>(orientation_j), this->m_params[typ_j]);
+
+                        #ifdef ENABLE_TBB
+                        thread_counters.local().overlap_checks++;
+                        #else
+                        counters.overlap_checks++;
+                        #endif
+
+                        unsigned int err = 0;
+
+                        if (m_quermass)
+                            {
+                            // check triple overlap of circumspheres
+
+                            // check triple overlap with i at old position
+                            vec3<Scalar> r_ij = vec3<Scalar>(postype_j) - pos_i_old - this->m_image_list[image_i[m]];
+
+                            // need to enable later when we have a better way of excluding particles from the image list calculation
+                            bool circumsphere_overlap = true || (h_overlaps[this->m_overlap_idx(type,typ_i)] && h_overlaps[this->m_overlap_idx(type,typ_j)]);
+                            circumsphere_overlap = circumsphere_overlap && check_circumsphere_overlap_three(shape_old, shape_j, shape_test,
+                                    r_ij, -r_jk+r_ij, m_sweep_radius, m_sweep_radius, 0.0);
+
+                            if (circumsphere_overlap
+                                && test_overlap_intersection(shape_old, shape_j, shape_test, r_ij, -r_jk+r_ij, err, m_sweep_radius, m_sweep_radius, 0.0))
+                                in_intersection_volume = true;
+
+                            if (in_intersection_volume)
+                                {
+                                // check triple overlap with i at new position
+                                r_ij = vec3<Scalar>(postype_j) - pos_i - this->m_image_list[image_i[m]];
+                                r_jk = ((i == j) ? pos_i : vec3<Scalar>(h_postype[j])) - pos_test - this->m_image_list[image_i[m]];
+
+                                // need to enable later when we have a better way of excluding particles from the image list calculation
+                                bool circumsphere_overlap = true || (h_overlaps[this->m_overlap_idx(type,typ_i)] && h_overlaps[this->m_overlap_idx(type,typ_j)]);
+                                circumsphere_overlap = circumsphere_overlap && check_circumsphere_overlap_three(shape_i, shape_j, shape_test, r_ij, -r_jk+r_ij,
+                                    m_sweep_radius, m_sweep_radius, 0.0);
+
+                                if (circumsphere_overlap
+                                    && test_overlap_intersection(shape_i, (i == j) ? shape_i : shape_j, shape_test, r_ij, -r_jk+r_ij, err,
+                                        m_sweep_radius, m_sweep_radius, 0.0))
+                                    in_intersection_volume = false;
+                                }
+                            }
+                        else
+                            {
+                            // check circumsphere overlap
+                            OverlapReal rsq = dot(r_jk,r_jk);
+                            OverlapReal DaDb = shape_test.getCircumsphereDiameter() + shape_j.getCircumsphereDiameter() + Scalar(2.0)*m_sweep_radius;
+                            bool circumsphere_overlap = (rsq*OverlapReal(4.0) <= DaDb * DaDb);
+
+                            if (h_overlaps[this->m_overlap_idx(type,typ_j)]
+                                && circumsphere_overlap
+                                && test_overlap(r_jk, shape_test, shape_j, err, 0.0, m_sweep_radius))
+                                {
+                                in_intersection_volume = true;
+                                }
+                            }
+
+                        if (err)
+                        #ifdef ENABLE_TBB
+                            thread_counters.local().overlap_err_count++;
+                        #else
+                            counters.overlap_err_count++;
+                        #endif
+
+                        if (in_intersection_volume)
+                            break;
+                        } // end loop over intersections
+
+                    // if not part of overlap volume in new config, reject
+                    if (in_intersection_volume)
+                        {
+                        accept = false;
+                        #ifndef ENABLE_TBB
+                        break;
+                        #else
+                        throw false;
+                        #endif
+                        }
+                    } // end loop over depletants
+                #ifdef ENABLE_TBB
+                    });
+                #endif
+
+                #ifndef ENABLE_TBB
+                if (!accept)
+                    break;
+                #endif
+                } // end loop over overlapping spheres
+            #ifdef ENABLE_TBB
+                });
+            #endif
+            }
+        // Depletant check for negative fugacity
+        else if (accept && m_fugacity[type] < 0.0)
+            {
+            Shape tmp(quat<Scalar>(), this->m_params[type]);
+            Scalar d_dep = tmp.getCircumsphereDiameter();
+
+            // find neighbors whose circumspheres overlap particle i's excluded volume circumsphere in the new configuration
+            Scalar range = m_quermass ? Scalar(2.0)*m_sweep_radius : d_dep;
+
+            // get new AABB and extend
+            vec3<Scalar> lower = aabb_i_local.getLower();
+            vec3<Scalar> upper = aabb_i_local.getUpper();
+            lower.x -= range; lower.y -= range; lower.z -= range;
+            upper.x += range; upper.y += range; upper.z += range;
+            detail::AABB aabb_local = detail::AABB(lower,upper);
+
+            // All image boxes (including the primary)
+            for (unsigned int cur_image = 0; cur_image < n_images; cur_image++)
+                {
+                vec3<Scalar> pos_i_image = pos_i + this->m_image_list[cur_image];
+                detail::AABB aabb = aabb_local;
+                aabb.translate(pos_i_image);
+
+                // stackless search
+                for (unsigned int cur_node_idx = 0; cur_node_idx < this->m_aabb_tree.getNumNodes(); cur_node_idx++)
+                    {
+                    if (detail::overlap(this->m_aabb_tree.getNodeAABB(cur_node_idx), aabb))
+                        {
+                        if (this->m_aabb_tree.isNodeLeaf(cur_node_idx))
+                            {
+                            for (unsigned int cur_p = 0; cur_p < this->m_aabb_tree.getNodeNumParticles(cur_node_idx); cur_p++)
+                                {
+                                // read in its position and orientation
+                                unsigned int j = this->m_aabb_tree.getNodeParticle(cur_node_idx, cur_p);
+
+                                unsigned int typ_j;
+                                vec3<Scalar> pos_j;
+                                if (i == j)
+                                    {
+                                    if (cur_image == 0)
+                                        continue;
+                                    else
+                                        {
+                                        typ_j = typ_i;
+                                        }
+                                    }
+                                else
+                                    {
+                                    // load the old position and orientation of the j particle
+                                    Scalar4 postype_j = h_postype[j];
+                                    pos_j = vec3<Scalar>(postype_j);
+                                    typ_j = __scalar_as_int(postype_j.w);
+                                    }
+
+                                vec3<Scalar> r_ij = pos_j - pos_i_image;
+                                Shape shape_j(quat<Scalar>(), this->m_params[typ_j]);
+
+                                if (shape_j.hasOrientation())
+                                    {
+                                    if (i == j)
+                                        shape_j.orientation = shape_i.orientation;
+                                    else
+                                        shape_j.orientation = quat<Scalar>(h_orientation[j]);
+                                    }
+
+                                // check excluded volume overlap
+                                detail::AABB aabb_j = shape_j.getAABB(pos_j);
+                                bool overlap_excluded = detail::overlap(aabb,aabb_j);
+
+                                // currently ignore overlap flags with quermass==true
+                                overlap_excluded = overlap_excluded && (m_quermass || h_overlaps[this->m_overlap_idx(type,typ_j)]);
+
+                                if (Shape::supportsSweepRadius() && overlap_excluded)
+                                    {
+                                    // do i and j overlap with their excluded volumes in the new configuration?
+                                    #ifdef ENABLE_TBB
+                                    thread_counters.local().overlap_checks++;
+                                    #else
+                                    counters.overlap_checks++;
+                                    #endif
+
+                                    unsigned int err = 0;
+                                    Scalar sweep_radius = m_quermass ? m_sweep_radius : Scalar(0.5)*tmp.getCircumsphereDiameter();
+                                    overlap_excluded = test_overlap(r_ij, shape_i, shape_j, err, sweep_radius, sweep_radius);
+
+                                    if (err)
+                                    #ifdef ENABLE_TBB
+                                        thread_counters.local().overlap_err_count++;
+                                    #else
+                                        counters.overlap_err_count++;
+                                    #endif
+                                    }
+                                if (overlap_excluded)
+                                    {
+                                    intersect_i.push_back(j);
+                                    image_i.push_back(cur_image);
+
+                                    // cache the translated AABB of particle j
+                                    aabb_j.translate(-this->m_image_list[cur_image]);
+                                    aabbs_i.push_back(aabb_j);
+                                    }
+                                }
+                            }
+                        }
+                    else
+                        {
+                        // skip ahead
+                        cur_node_idx += this->m_aabb_tree.getNodeSkip(cur_node_idx);
+                        }
+                    }  // end loop over AABB nodes
+                } // end loop over images
+
+
+            // now, we have a list of intersecting spheres, sample in the union of intersection volumes
+            // we sample from their union by checking if any generated position falls in the intersection
+            // of their AABBs and if so, only accepting it if it was generated from neighbor j_min
+
+            // world AABB of particle i
+            detail::AABB aabb_i = aabb_i_local;
+            aabb_i.translate(pos_i);
+
+            // extend AABB i by sweep radius
+            range = Scalar(0.5)*d_dep+m_sweep_radius;
+            vec3<Scalar> lower_i = aabb_i.getLower();
+            vec3<Scalar> upper_i = aabb_i.getUpper();
+            lower_i.x -= range; lower_i.y -= range; lower_i.z -= range;
+            upper_i.x += range; upper_i.y += range; upper_i.z += range;
+
+            // for every pairwise intersection
+            #ifdef ENABLE_TBB
+            tbb::parallel_for(tbb::blocked_range<unsigned int>(0, (unsigned int)intersect_i.size()),
+                [=, &intersect_i, &image_i, &aabbs_i,
+                    &accept, &rng_depletants_parallel,
+                    &thread_counters, &thread_implicit_counters](const tbb::blocked_range<unsigned int>& s) {
+            for (unsigned int k = s.begin(); k != s.end(); ++k)
+            #else
+            for (unsigned int k = 0; k < intersect_i.size(); ++k)
+            #endif
+                {
+                // world AABB of shape j, in image of i
+                detail::AABB aabb_j = aabbs_i[k];
+
+                // extend AABB j by sweep radius
+                vec3<Scalar> lower_j = aabb_j.getLower();
+                vec3<Scalar> upper_j = aabb_j.getUpper();
+                lower_j.x -= range; lower_j.y -= range; lower_j.z -= range;
+                upper_j.x += range; upper_j.y += range; upper_j.z += range;
+
+                // we already know the AABBs are overlapping, compute their intersection
+                vec3<Scalar> intersect_lower, intersect_upper;
+                intersect_lower.x = std::max(lower_i.x, lower_j.x);
+                intersect_lower.y = std::max(lower_i.y, lower_j.y);
+                intersect_lower.z = std::max(lower_i.z, lower_j.z);
+                intersect_upper.x = std::min(upper_i.x, upper_j.x);
+                intersect_upper.y = std::min(upper_i.y, upper_j.y);
+                intersect_upper.z = std::min(upper_i.z, upper_j.z);
+
+                detail::AABB aabb_intersect(intersect_lower, intersect_upper);
+
+                // intersection AABB volume
+                Scalar V = (intersect_upper.x-intersect_lower.x)*(intersect_upper.y-intersect_lower.y)*(intersect_upper.z-intersect_lower.z);
+
+                // chooose the number of depletants in the intersection volume
+                hoomd::PoissonDistribution<Scalar> poisson(-m_fugacity[type]*V);
+                #ifdef ENABLE_TBB
+                hoomd::RandomGenerator& my_rng = rng_depletants_parallel.local();
+                #else
+                hoomd::RandomGenerator& my_rng = rng_depletants;
+                #endif
+
+                unsigned int n = poisson(my_rng);
+
+                // for every depletant
+                #ifdef ENABLE_TBB
+                tbb::parallel_for(tbb::blocked_range<unsigned int>(0, (unsigned int)n),
+                    [=, &intersect_i, &image_i, &aabbs_i,
+                        &accept, &rng_depletants_parallel,
+                        &thread_counters, &thread_implicit_counters](const tbb::blocked_range<unsigned int>& t) {
+                for (unsigned int l = t.begin(); l != t.end(); ++l)
+                #else
+                for (unsigned int l = 0; l < n; ++l)
+                #endif
+                    {
+                    #ifdef ENABLE_TBB
+                    hoomd::RandomGenerator& my_rng = rng_depletants_parallel.local();
+                    #else
+                    hoomd::RandomGenerator& my_rng = rng_depletants;
+                    #endif
+
+                    #ifdef ENABLE_TBB
+                    thread_implicit_counters[type].local().insert_count++;
+                    #else
+                    implicit_counters[type].insert_count++;
+                    #endif
+
+                    vec3<Scalar> pos_test = generatePositionInAABB(my_rng, aabb_intersect);
+                    Shape shape_test(quat<Scalar>(), this->m_params[type]);
+                    if (shape_test.hasOrientation())
+                        {
+                        shape_test.orientation = generateRandomOrientation(my_rng);
+                        }
+
+                    // check if depletant falls in other intersection volumes (new)
+                    bool active = true;
+
+                    // check against any other AABB preceding this
+
+                    // load the depletant's AABB and extend it
+                    detail::AABB aabb_test = shape_test.getAABB(pos_test);
+                    vec3<Scalar> lower_test = aabb_test.getLower();
+                    vec3<Scalar> upper_test = aabb_test.getUpper();
+                    lower_test.x -= m_sweep_radius; lower_test.y -= m_sweep_radius; lower_test.z -= m_sweep_radius;
+                    upper_test.x += m_sweep_radius; upper_test.y += m_sweep_radius; upper_test.z += m_sweep_radius;
+                    aabb_test = detail::AABB(lower_test, upper_test);
+
+                    for (unsigned int m = 0; m < k; ++m)
+                        {
+                        if (detail::overlap(aabb_test,aabbs_i[m]))
+                            {
+                            active = false;
+                            break;
+                            }
+                        }
+
+                    if (!active)
+                        continue;
+
+                    // depletant falls in intersection volume between circumspheres
+
+                    if (!m_quermass)
+                        {
+                        // Check if the new configuration of particle i generates an overlap
+                        bool overlap_new = false;
+
+                            {
+                            vec3<Scalar> r_ij = pos_i - pos_test;
+
+                            OverlapReal rsq = dot(r_ij,r_ij);
+                            OverlapReal DaDb = shape_test.getCircumsphereDiameter() + shape_i.getCircumsphereDiameter() + OverlapReal(2.0)*m_sweep_radius;
+                            bool circumsphere_overlap = (rsq*OverlapReal(4.0) <= DaDb * DaDb);
+
+                            if (m_quermass || h_overlaps[this->m_overlap_idx(type, typ_i)])
+                                {
+                                #ifdef ENABLE_TBB
+                                thread_counters.local().overlap_checks++;
+                                #else
+                                counters.overlap_checks++;
+                                #endif
+
+                                unsigned int err = 0;
+                                if (circumsphere_overlap && test_overlap(r_ij, shape_test, shape_i, err, 0.0, m_sweep_radius))
+                                    {
+                                    overlap_new = true;
+                                    }
+                                if (err)
+                                #ifdef ENABLE_TBB
+                                    thread_counters.local().overlap_err_count++;
+                                #else
+                                    counters.overlap_err_count++;
+                                #endif
+                                }
+                            }
+
+                        if (!overlap_new)
+                            continue;
+
+                        // Check if the old configuration of particle i generates an overlap
+                        bool overlap_old = false;
+
+                        vec3<Scalar> r_ij = vec3<Scalar>(h_postype[i]) - pos_test;
+
+                        OverlapReal rsq = dot(r_ij,r_ij);
+                        OverlapReal DaDb = shape_test.getCircumsphereDiameter() + shape_old.getCircumsphereDiameter() + OverlapReal(2.0)*m_sweep_radius;
+                        bool circumsphere_overlap = (rsq*OverlapReal(4.0) <= DaDb * DaDb);
+
+                        if (m_quermass || h_overlaps[this->m_overlap_idx(type, typ_i)])
+                            {
+                            #ifdef ENABLE_TBB
+                            thread_counters.local().overlap_checks++;
+                            #else
+                            counters.overlap_checks++;
+                            #endif
+
+                            unsigned int err = 0;
+                            if (circumsphere_overlap && test_overlap(r_ij, shape_test, shape_old, err, 0.0, m_sweep_radius))
+                                {
+                                overlap_old = true;
+                                }
+                            if (err)
+                            #ifdef ENABLE_TBB
+                                thread_counters.local().overlap_err_count++;
+                            #else
+                                counters.overlap_err_count++;
+                            #endif
+                            }
+
+                        if (overlap_old)
+                            {
+                            // all is ok
+                            continue;
+                            }
+                        }
+
+                    bool in_new_intersection_volume = false;
+
+                    vec3<Scalar> pos_i_old(h_postype[i]);
+
+                    for (unsigned int m = 0; m < intersect_i.size(); ++m)
+                        {
+                        // read in its position and orientation
+                        unsigned int j = intersect_i[m];
+
+                        vec3<Scalar> r_jk = ((i == j) ? pos_i : vec3<Scalar>(h_postype[j])) - pos_test -
+                            this->m_image_list[image_i[m]];
+
+                        unsigned int typ_j = (i == j) ? typ_i : __scalar_as_int(h_postype[j].w);
+                        Shape shape_j((i == j) ? shape_i.orientation : quat<Scalar>(h_orientation[j]), this->m_params[typ_j]);
+
+                        #ifdef ENABLE_TBB
+                        thread_counters.local().overlap_checks++;
+                        #else
+                        counters.overlap_checks++;
+                        #endif
+
+                        unsigned int err = 0;
+                        if (m_quermass)
+                            {
+                            // check triple overlap of circumspheres
+                            vec3<Scalar> r_ij = ( (i==j) ? pos_i : vec3<Scalar>(h_postype[j])) - pos_i -
+                                this->m_image_list[image_i[m]];
+
+                            // need to enable later when we have a better way of excluding particles from the image list calculation
+                            bool circumsphere_overlap = true || (h_overlaps[this->m_overlap_idx(type,typ_i)] && h_overlaps[this->m_overlap_idx(type,typ_j)]);
+                            circumsphere_overlap = circumsphere_overlap && check_circumsphere_overlap_three(shape_i, shape_j, shape_test, r_ij, r_ij - r_jk,
+                                m_sweep_radius, m_sweep_radius, 0.0);
+
+                            // check triple overlap with new configuration
+                            unsigned int err = 0;
+
+                            if (circumsphere_overlap
+                                && test_overlap_intersection(shape_i, shape_j, shape_test, r_ij, r_ij - r_jk, err, m_sweep_radius, m_sweep_radius, 0.0))
+                                in_new_intersection_volume = true;
+
+                            if (in_new_intersection_volume)
+                                {
+                                // check triple overlap with old configuration
+                                r_ij = vec3<Scalar>(h_postype[j]) - pos_i_old - this->m_image_list[image_i[m]];
+                                r_jk = vec3<Scalar>(h_postype[j]) - pos_test - this->m_image_list[image_i[m]];
+
+                                // need to enable later when we have a better way of excluding particles from the image list calculation
+                                bool circumsphere_overlap = true || (h_overlaps[this->m_overlap_idx(type,typ_i)] && h_overlaps[this->m_overlap_idx(type,typ_j)]);
+                                circumsphere_overlap = circumsphere_overlap && check_circumsphere_overlap_three(shape_old, shape_j, shape_test, r_ij, r_ij - r_jk,
+                                    m_sweep_radius, m_sweep_radius, 0.0);
+
+                                if (circumsphere_overlap
+                                    && test_overlap_intersection(shape_old, (i == j) ? shape_old : shape_j, shape_test, r_ij, r_ij - r_jk, err,
+                                        m_sweep_radius, m_sweep_radius, 0.0))
+                                    in_new_intersection_volume = false;
+                                }
+                            if (err)
+                            #ifdef ENABLE_TBB
+                                thread_counters.local().overlap_err_count++;
+                            #else
+                                counters.overlap_err_count++;
+                            #endif
+                            }
+                        else
+                            {
+                            // check circumsphere overlap
+                            OverlapReal rsq = dot(r_jk,r_jk);
+                            OverlapReal DaDb = shape_test.getCircumsphereDiameter() + shape_j.getCircumsphereDiameter() + OverlapReal(2.0)*m_sweep_radius;
+                            bool circumsphere_overlap = (rsq*OverlapReal(4.0) <= DaDb * DaDb);
+
+                            if (h_overlaps[this->m_overlap_idx(type,typ_j)]
+                                && circumsphere_overlap
+                                && test_overlap(r_jk, shape_test, shape_j, err, 0.0, m_sweep_radius))
+                                {
+                                in_new_intersection_volume = true;
+                                }
+                            if (err)
+                            #ifdef ENABLE_TBB
+                                thread_counters.local().overlap_err_count++;
+                            #else
+                                counters.overlap_err_count++;
+                            #endif
+                            }
+
+                        if (in_new_intersection_volume)
+                            break;
+                        }
+
+                    if (in_new_intersection_volume)
+                        {
+                        accept = false;
+
+                        #ifndef ENABLE_TBB
+                        // early exit
+                        break;
+                        #else
+                        throw false;
+                        #endif
+                        }
+                    } // end loop over depletants
+                #ifdef ENABLE_TBB
+                    });
+                #endif
+
+                #ifndef ENABLE_TBB
+                if (!accept) break;
+                #endif
+                } // end loop over overlapping spheres
+            #ifdef ENABLE_TBB
+                });
+            #endif
+            } // end depletant placement
+        }
+    #ifdef ENABLE_TBB
+    } catch (bool b) { }
+    #endif
+
+    #ifdef ENABLE_TBB
+    // reduce counters
+    for (auto i = thread_counters.begin(); i != thread_counters.end(); ++i)
+        {
+        counters = counters + *i;
+        }
+
+    for (unsigned int i = 0; i < this->m_pdata->getNTypes(); ++i)
+        for (auto it= thread_implicit_counters[i].begin(); it != thread_implicit_counters[i].end(); ++it)
+            {
+            implicit_counters[i] = implicit_counters[i] + *it;
+            }
+    #endif
+
+    return accept;
+    }
+
 //! Export the IntegratorHPMCMono class to python
 /*! \param name Name of the class in the exported python module
     \tparam Shape An instantiation of IntegratorHPMCMono<Shape> will be exported
@@ -1820,7 +2984,22 @@ template < class Shape > void export_IntegratorHPMCMono(pybind11::module& m, con
           .def("connectGSDSignal", &IntegratorHPMCMono<Shape>::connectGSDSignal)
           .def("restoreStateGSD", &IntegratorHPMCMono<Shape>::restoreStateGSD)
           .def("py_test_overlap", &IntegratorHPMCMono<Shape>::py_test_overlap)
+          .def("setDepletantFugacity", &IntegratorHPMCMono<Shape>::setDepletantFugacity)
+          .def("getImplicitCounters", &IntegratorHPMCMono<Shape>::getImplicitCounters)
+          .def("getDepletantFugacity", &IntegratorHPMCMono<Shape>::getDepletantFugacity)
+          .def("setQuermassMode", &IntegratorHPMCMono<Shape>::setQuermassMode)
+          .def("setSweepRadius", &IntegratorHPMCMono<Shape>::setSweepRadius)
+          .def("getQuermassMode", &IntegratorHPMCMono<Shape>::getQuermassMode)
+          .def("getSweepRadius", &IntegratorHPMCMono<Shape>::getSweepRadius)
           ;
+    }
+
+//! Export the counters for depletants
+inline void export_hpmc_implicit_counters(pybind11::module& m)
+    {
+    pybind11::class_< hpmc_implicit_counters_t >(m, "hpmc_implicit_counters_t")
+    .def_readwrite("insert_count", &hpmc_implicit_counters_t::insert_count)
+    ;
     }
 
 } // end namespace hpmc
